+++ example/package/test/integration_test.dart
@@ -16,7 +16,7 @@ Future<void> main() async {
   final testData =
       Platform.environment['API_TEST_DATA_FILE'] ?? 'api_test_data.json';
   final data = await File(testData).readAsString();
-  final json = jsonDecode(data);
+  final json = jsonDecode(data.injectEnvironmentVariables());
 
   final baseUrl = json['server'] as String;
   final operations = json['operations'] as List;
@@ -57,6 +57,20 @@ Api createApi({
   return Api(
     client: dio,
   );
+}
+
+extension EnvironmentVariableTemplateExtentions on String {
+  String injectEnvironmentVariables() {
+    var result = this;
+    final regexp = RegExp(r'\$\{\{([a-zA-Z\_\-0-9]+)\}\}');
+    RegExpMatch? match;
+    while ((match = regexp.firstMatch(result)) != null) {
+      final name = match!.group(1);
+      result = result.replaceRange(
+          match.start, match.end, Platform.environment[name] ?? '<$name?>');
+    }
+    return result;
+  }
 }
 
 extension ApiTestExtensions on Api {