+++ openapi_schema_generator/lib/src/generators/api.dart
@@ -227,6 +227,10 @@ class ApiGenerator {
       );
     }
 
+    code.writeln('options: Options('
+        'headers: { \'Content-Type\' : \'application/json\',}'
+        ',),');
+
     code.writeln(');');
 
     code.writeln('return $responseType(response);');