+++ openapi_schema_generator/lib/src/utilities/fix_trailing_comma.dart
@@ -1,6 +1,7 @@
 final _missingTrailingComma = RegExp(r'([^\,])\s*\}\)');
 
 extension FixTrailingCommas on String {
+  /// Adds missing commas to Dart generated code.
   String fixTrailingCommas() {
     var match = _missingTrailingComma.firstMatch(this);
     var result = this;